Renamed aggregator base controller to NestJS standard
