+++ src/modules/aggregator/controllers/aggregates/instance/aggregator.instance.weapon.aggregate.controller.ts
@@ -1,6 +1,6 @@
 import {BadRequestException, Controller} from '@nestjs/common';
 import {Ctx, MessagePattern, Payload, RmqContext} from '@nestjs/microservices';
-import AggregatorBaseController from '../../AggregatorBaseController';
+import AggregatorBaseController from '../../aggregator.base.controller';
 import {MQAcceptedPatterns} from '../../../../data/constants/MQAcceptedPatterns';
 import AggregatorMessageInterface from '../../../interfaces/AggregatorMessageInterface';
 import InstanceWeaponAggregateEntity from '../../../../data/entities/aggregate/instance/instance.weapon.aggregate.entity';
