+++ src/modules/aggregator/controllers/events/aggregator.instance.death.event.controller.ts
@@ -2,7 +2,7 @@ import {BadRequestException, Controller} from '@nestjs/common';
 import {Ctx, MessagePattern, Payload, RmqContext} from '@nestjs/microservices';
 import InstanceDeathEntity from '../../../data/entities/instance/instance.death.entity';
 import AggregatorMessageInterface from '../../interfaces/AggregatorMessageInterface';
-import AggregatorBaseController from '../AggregatorBaseController';
+import AggregatorBaseController from '../aggregator.base.controller';
 import {MQAcceptedPatterns} from '../../../data/constants/MQAcceptedPatterns';
 
 @Controller()
