+++ src/modules/aggregator/controllers/events/aggregator.instance.facility.control.event.controller.ts
@@ -1,7 +1,7 @@
 import {BadRequestException, Controller} from '@nestjs/common';
 import {Ctx, MessagePattern, Payload, RmqContext} from '@nestjs/microservices';
 import AggregatorMessageInterface from '../../interfaces/AggregatorMessageInterface';
-import AggregatorBaseController from '../AggregatorBaseController';
+import AggregatorBaseController from '../aggregator.base.controller';
 import {MQAcceptedPatterns} from '../../../data/constants/MQAcceptedPatterns';
 import InstanceFacilityControlEntity from '../../../data/entities/instance/instance.facilitycontrol.entity';
